+++ R/loglikelihood.R
@@ -524,9 +524,11 @@ get_alpha_mt <- function(data, Y2, p, M, d,
                          weight_function=c("relative_dens", "logistic", "mlogit", "exponential", "threshold", "exogenous"),
                          weightfun_pars=NULL, all_A, all_boldA, all_Omegas, weightpars, all_mu, epsilon, log_mvdvalues=NULL) {
   weight_function <- match.arg(weight_function)
+
   if(weight_function == "exogenous") {
     return(weightfun_pars)
   }
+
   if(is.null(log_mvdvalues)) {
     T_obs <- ifelse(missing(data), 1, nrow(data) - p) # simulate.stvar and estim_LS use without data, needs to return 1 if M=1.
     if(M == 1) {