+++ src/get_Bt_Cpp.cpp
@@ -7,7 +7,8 @@
 //'
 //' @description This internal function takes a cube of matrices (\code{all_Omegas}) and a matrix of weights (\code{alpha_mt}),
 //' and calculates the weighted sums of the matrices in the cube. For each row in \code{alpha_mt}, it computes
-//' a weighted sum of matrices, and returns the
+//' a weighted sum of matrices, and returns the (3D array) such that each slice contains the weighted sum of the matrices.
+//' Note that the argument \code{all_Omegas} should contain the impact matrices of the regimes (and not the covariance matrices).
 //'
 //' @inheritParams check_Bt_Cpp
 //'